modify Ui code

+++ src/app/pages/admin/userlist/userlist.component.ts
@@ -1,8 +1,7 @@
 import { Component } from "@angular/core";
-import { MatPaginator, MatPaginatorModule } from "@angular/material/paginator";
 import { MatSnackBar } from "@angular/material/snack-bar";
-import { MatSort, MatSortModule } from "@angular/material/sort";
 import { AuthuserService } from "src/app/services/authuser.service";
+import Swal from "sweetalert2";
 
 @Component({
   selector: "app-userlist",
@@ -10,10 +9,13 @@ import { AuthuserService } from "src/app/services/authuser.service";
   styleUrls: ["./userlist.component.css"],
 })
 export class UserlistComponent {
-  constructor(private userservice: AuthuserService,private snackBar:MatSnackBar) {}
+  constructor(
+    private userservice: AuthuserService,
+    private snackBar: MatSnackBar
+  ) {}
 
-  ngOnInit(){
-    this.getUser()
+  ngOnInit() {
+    this.getUser();
   }
   data: any = [];
   getUser() {
@@ -24,18 +26,32 @@ export class UserlistComponent {
     });
   }
 
-  deleteUser(id:number){
-    console.log("user deleted successfully")
-    console.log("11111",id)
-    this.userservice.deleteUser(id).subscribe((res)=>{
-      console.log("45468468",res)
-      this.snackBar.open('User deleted successfully', 'Close', {
-        duration: 2000,
-        horizontalPosition: 'center',
-        verticalPosition: 'top',
-      });
-      this.getUser();
+  deleteUser(id: number) {
+    console.log("user deleted successfully");
+    console.log("11111", id);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't delete this user!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Deleted!",
+          text: "User has been deleted successfully.",
+          icon: "success",
+        });
+        this.userservice.deleteUser(id).subscribe((res) => {
+          console.log("45468468", res);
+          this.getUser();
+        });
+      }else{
+        console.log("User deletion is failed")
+      }
 
-    })
+    });
   }
 }
