+++ src/app/pages/admin/view-quizzes/view-quizzes.component.ts
@@ -1,8 +1,8 @@
 import { Component } from "@angular/core";
 import { Router } from "@angular/router";
 import { QuizzesService } from "src/app/services/quizzes.service";
-import { MatSnackBar } from '@angular/material/snack-bar';
-
+import { MatSnackBar } from "@angular/material/snack-bar";
+import Swal from "sweetalert2";
 
 @Component({
   selector: "app-view-quizzes",
@@ -10,7 +10,11 @@ import { MatSnackBar } from '@angular/material/snack-bar';
   styleUrls: ["./view-quizzes.component.css"],
 })
 export class ViewQuizzesComponent {
-  constructor(private quizservice: QuizzesService, private router:Router,private snackBar: MatSnackBar) {}
+  constructor(
+    private quizservice: QuizzesService,
+    private router: Router,
+    private snackBar: MatSnackBar
+  ) {}
 
   ngOnInit() {
     this.getQuizzes();
@@ -25,22 +29,37 @@ export class ViewQuizzesComponent {
   }
 
   deleteQuiz(id: number) {
-    this.quizservice.deleteQuizzes(id).subscribe((res) => {
-      console.log("first", res);
-      console.log("quiz deleted ");
-      this.snackBar.open('Quiz deleted successfully', 'Close', {
-        duration: 2000,
-        horizontalPosition: 'center',
-        verticalPosition: 'top',
-      });
-      this.getQuizzes();
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't delete this Quiz!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Deleted!",
+          text: "Quiz has been deleted successfully.",
+          icon: "success",
+        });
+        this.quizservice.deleteQuizzes(id).subscribe((res) => {
+          console.log("first", res);
+          console.log("quiz deleted ");
+
+          this.getQuizzes();
+        });
+      } else {
+        console.log("quiz deletion is failed");
+      }
     });
   }
 
-  navigate(id:number){
-    this.router.navigate(['/admin/updatequiz', id]);
+  navigate(id: number) {
+    this.router.navigate(["/admin/updatequiz", id]);
   }
-  navigateView(id:number){
-    this.router.navigate(['/admin/viewquestion',id])
+  navigateView(id: number) {
+    this.router.navigate(["/admin/viewquestion", id]);
   }
 }
